requirements.txt
streamlit
streamlit-drawable-canvas
numpy
pandas
streamlit-javascript
google-generativeai
python-dotenv